+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [21.12.0] - Aspose Words Cloud for Dart 21.12 Release Notes
+
+- Added 'timeout' parameter to api configuration
+
 ## [21.11.0] - Aspose Words Cloud for Dart 21.11 Release Notes
 
 - Support encryption of 'CommonRequest.Password'. Automatic encryption of all passwords sent to the API server as request parameters.